Create TastyBuds.sql
Delve into the sweet world of data from an Ice cream shop with this immersive exploration. Uncover the flavors that reign supreme, the busiest days and time of the day, the demographics of delighted customers and the total revenue as days turn into weeks and weeks into months. this data journey promises to serve up insights as refreshing as a scoop of your favourite flavour.